Make support for HTTP protocol non-mandatory
HTTP protocol support in libjson-rpc-cpp relies on libcurl which
may not be a desirable dependency for some embedded systems. If
another transport for JSON-RPC is used, support for HTTP could be
switched off. This patch makes support for HTTP an option that is
ON by default but could be switched off if not desired.